# improve performance deepcopying DependencyDescriptors