On a large workspace with ~300 pkgs this patch reduces the time for `colcon list` from ~1.0s to ~0.9s.

Profiling the `get_recursive_dependencies()` function shows that from ~370ms before ~310ms where spent in `deepcopy()`. With this change the `deepcopy()` part is reduced to 110ms.